Move settings => setting.
